refactor: #148 Add review feedback
