--- packages/liferay-theme-tasks/index.js
@@ -1,6 +1,0 @@
-/**
- * SPDX-FileCopyrightText: © 2017 Liferay, Inc. <https://liferay.com>
- * SPDX-License-Identifier: MIT
- */
-
-module.exports = require('./theme');
