+++ packages/liferay-theme-tasks/lib/project/store.js
