+++ packages/liferay-theme-tasks/lib/project/theme-config.js
